refactior: api 정리
(cherry picked from commit a6704b7844818775f01ed769c8d7ec9795d03761)
