+++ app/mydashboard/_components/InvitationList/index.tsx
@@ -6,18 +6,15 @@ import { useQuery } from '@tanstack/react-query';
 import Image from 'next/image';
 import styled from 'styled-components';
 
+import { getInitialInvitionList, Invitation } from '@apis/invitations/getInitialInvitionList';
 import searchIcon from '@public/images/icons/search-filledBlack-333236-w22-h22.svg';
 import uninvitedSvg from '@public/images/logos/unInvited_filledGray_D9D9D9-w100-h100.svg';
 import { mediaBreakpoint } from '@styles/mediaBreakpoint';
 
 import InvitationItem from './invitationItem';
-import {
-  getInitialInvitionList,
-  getMoreInvitionList,
-  getSearchedInvitationList,
-  Invitation,
-  putInvitationAnswer,
-} from '../apis/api';
+import { getMoreInvitionList } from '../../../../apis/invitations/getMoreInvitionList';
+import { getSearchedInvitationList } from '../../../../apis/invitations/getSearchedInvitationList';
+import { putInvitationAnswer } from '../../../../apis/invitations/putInvitationAnswer';
 import InvitationText from '../commons/InvitationText';
 
 export default function InvitationList() {
