--- app/mydashboard/_components/utils/index.ts
@@ -1,6 +1,0 @@
-import { putInvitationAnswer } from '../apis/api';
-
-// 초대 답변
-export const handleInvitationAnswer = async (invitationId: number, answer: boolean) => {
-  await putInvitationAnswer(invitationId, answer);
-};
